% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{add_cardinality_constraint}
\alias{add_cardinality_constraint}
\title{Add Cardinality Constraint to Constraints Object}
\usage{
add_cardinality_constraint(constraints, min = 0, max = NULL)
}
\arguments{
\item{constraints}{constraints object}

\item{min}{minimum constraint value. inclusive}

\item{max}{maximum constraint value. inclusive}
}
\value{
updated constraints object
}
\description{
Add Cardinality Constraint to Constraints Object
}
