% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{add_cash_constraint}
\alias{add_cash_constraint}
\title{Add Cash Constraint to Constraints Object}
\usage{
add_cash_constraint(constraints, min = 0, max = 1)
}
\arguments{
\item{constraints}{constraints object}

\item{min}{minimum constraint value. inclusive}

\item{max}{maximum constraint value. inclusive}
}
\value{
updated constraints object
}
\description{
Cash constraints constrain the share of a portfolio's cash position
}
