% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{add_constraint}
\alias{add_constraint}
\title{Add Constraint to Constraints Object}
\usage{
add_constraint(constraints, constraint)
}
\arguments{
\item{constraints}{constraints object}

\item{constraint}{constraint object}
}
\value{
updated constraints object
}
\description{
Appends constraint object to constraints list of contraints. Adds id field to
constraint object
}
