% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{add_group_constraint}
\alias{add_group_constraint}
\title{Add Group Constraint to Constraints Object}
\usage{
add_group_constraint(constraints, symbols = NULL, min = 0, max = 1)
}
\arguments{
\item{constraints}{constraints object}

\item{symbols}{1 or more symbols}

\item{min}{minimum constraint value. inclusive}

\item{max}{maximum constraint value. inclusive}
}
\value{
updated constraints object
}
\description{
Add Group Constraint to Constraints Object
}
