% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{add_min_return}
\alias{add_min_return}
\title{Add Minimum Return Performance Constraint to Contraints Object}
\usage{
add_min_return(cobj, min = NULL)
}
\arguments{
\item{cobj}{constraints object}

\item{min}{minumum return value}
}
\value{
updated constraints object
}
\description{
Add Minimum Return Performance Constraint to Contraints Object
}
