% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{add_mu}
\alias{add_mu}
\title{Add Mu Estimates}
\usage{
add_mu(eobj, mu)
}
\arguments{
\item{eobj}{estimates object}

\item{mu}{data.frame with mu estimates. requires symbol and return column
names}
}
\value{
upated estimates object
}
\description{
Function adds pre-calculated mu estimates. Mu estimates should contain
estimates for all symbols in estimate object
}
