% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{add_sample_mu}
\alias{add_sample_mu}
\title{Add Calculated Sample Mu to Estimates}
\usage{
add_sample_mu(eobj, fun = "mean", ...)
}
\arguments{
\item{eobj}{estimates object}

\item{fun}{summarise function. default is mean}

\item{...}{additional arguments to pass to function}
}
\value{
updated estimates object
}
\description{
Function adds mu estimates. Allows for configurable function applied to
sample returns
}
