% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{add_sample_sigma}
\alias{add_sample_sigma}
\title{Add Calculated Sample Sigma to Estimates}
\usage{
add_sample_sigma(eobj, fun = "cov", ...)
}
\arguments{
\item{eobj}{estimates object}

\item{fun}{summarise function. default is cov}

\item{...}{additional arguments to pass to function}
}
\value{
updated estimates object
}
\description{
Function adds mu estimates. Allows for configurable function applied to
sample returns
}
