% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{add_shrink_sigma}
\alias{add_shrink_sigma}
\title{Add Calculated Sample Sigma to Estimates}
\usage{
add_shrink_sigma(eobj, ...)
}
\arguments{
\item{eobj}{estimates object}

\item{...}{additional arguments to pass to function}
}
\value{
updated estimates object
}
\description{
Functions adds shrunk sigma estimates using lediot and wolf shrinkage
estimator appraoch. See \code{\link[corpcor]{cov.shrink}}
}
\details{
Default behavior is to estimate lambda and lambda.var empricially. These can
lead to strong shrinkage estimates. Manual values can be provided to both
ranges between [0,1]
}
