% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{add_sigma}
\alias{add_sigma}
\title{Add Sigma Estimates}
\usage{
add_sigma(eobj, sigma)
}
\arguments{
\item{eobj}{estimates object}

\item{sigma}{sigma estimate matrix}
}
\description{
Function to add pre-calculated sigma matrix to estimates
}
