% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{add_symbol_constraint}
\alias{add_symbol_constraint}
\title{Add Symbol Constraint to Constraints Object}
\usage{
add_symbol_constraint(constraints, symbol = NULL, min = 0, max = 1)
}
\arguments{
\item{constraints}{constraints object}

\item{symbol}{single symbol to constrain}

\item{min}{minimum constraint value. inclusive}

\item{max}{maximum constraint value. inclusive}
}
\value{
updated constraints object
}
\description{
Symbol constraints constrain the share of a portfolio's market value a symbol
can have
}
