% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{cardinality_constraint}
\alias{cardinality_constraint}
\title{Cardinality Constraint Constructer}
\usage{
cardinality_constraint(min, max)
}
\arguments{
\item{min}{minimum constraint value. inclusive}

\item{max}{maximum constraint value. inclusive}
}
\value{
object of class cardinality_constraint
}
\description{
Cardinality constraints limit number of symbols portfolio can hold. Inherits
from constraint class
}
