% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{cash_constraint}
\alias{cash_constraint}
\title{Cash Constraint Constructer}
\usage{
cash_constraint(min, max)
}
\arguments{
\item{min}{minimum constraint value. inclusive}

\item{max}{maximum constraint value. inclusive}
}
\value{
cash constaint object
}
\description{
Inherits from constraint class
}
