% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{check_constraint}
\alias{check_constraint}
\alias{check_constraint.symbol_constraint}
\alias{check_constraint.cash_constraint}
\alias{check_constraint.cardinality_constraint}
\alias{check_constraint.group_constraint}
\alias{check_constraint.performance_constraint}
\title{Check Constraint}
\usage{
check_constraint(constraint, pobj = NULL, holdings = NULL,
  stats = NULL, ...)

\method{check_constraint}{symbol_constraint}(constraint, pobj = NULL,
  holdings, stats = NULL, ...)

\method{check_constraint}{cash_constraint}(constraint, pobj,
  holdings = NULL, stats = NULL, ...)

\method{check_constraint}{cardinality_constraint}(constraint,
  pobj = NULL, holdings, stats = NULL, ...)

\method{check_constraint}{group_constraint}(constraint, pobj = NULL,
  holdings, stats = NULL, ...)

\method{check_constraint}{performance_constraint}(constraint,
  pobj = NULL, holdings = NULL, stats, ...)
}
\arguments{
\item{constraint}{constraint object}

\item{pobj}{portfolio object}

\item{holdings}{portfolio holdings}

\item{stats}{portfolio statistics}

\item{...}{additional parameters. not currently implemented}
}
\value{
data.frame with summary of constraint check
}
\description{
Check portfolio's holdings and estimated statistics against constraint
}
