% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{check_constraints}
\alias{check_constraints}
\title{Check Constraints}
\usage{
check_constraints(constraints, portfolio, estimates)
}
\arguments{
\item{constraints}{constraints object}

\item{portfolio}{portfolio object}

\item{estimates}{estimates object}
}
\value{
data.frame with summary of constraint checks
}
\description{
Check the validity of portfolio against constraints.
}
\details{
Function applies all constraints on a portfolio holdings and estimates
}
