% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{constraint}
\alias{constraint}
\title{Constraint Constructer}
\usage{
constraint(type, args, min, max)
}
\arguments{
\item{type}{character value for type of constraint. valid types are symbol,
cardinality, group and performance}

\item{args}{symbol, symbols or statistic to test}

\item{min}{minimum constraint value. inclusive}

\item{max}{maximum constraint value. inclusive}
}
\value{
object of class constraint
}
\description{
Creates an object of class constraint
}
