% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{constraints}
\alias{constraints}
\title{Constraints Constructer}
\usage{
constraints(symbols, constraints = NULL)
}
\arguments{
\item{symbols}{vector of symbols to apply constraints to. Symbols constrain
tradable symbols in optimization}

\item{constraints}{list of contraints}
}
\value{
constraints object
}
\description{
Creates an object of class constraints
}
