% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{estimates}
\alias{estimates}
\title{Estimate Helper}
\usage{
estimates(symbols, start_date, end_date, grain, periods = 1,
  prices = NULL, returns = NULL)
}
\arguments{
\item{symbols}{vector of security symbols}

\item{start_date}{start date for historical prices}

\item{end_date}{end date for historical prices}

\item{grain}{time grain of returns. Can be either day, week, month, year}

\item{periods}{number of time periods}

\item{prices}{optional input for security prices}

\item{returns}{optional input for security returns}
}
\description{
Function to create a new estimates object. If prices not supplied, get_prices
function called and returns calculated. Alternatively, returns and prices can
be supplied.
}
