% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade-helpers.R
\name{execute_buy}
\alias{execute_buy}
\title{Execute Buy Trade Function}
\usage{
execute_buy(buy, pobj, prices)
}
\arguments{
\item{buy}{data.frame with buy trade information. Result of get_buy_trades
function}

\item{pobj}{portfolio object}

\item{prices}{current prices of symbols being traded}
}
\value{
updated portfolio object
}
\description{
Updates portfolio with buy trade provided
}
