% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade-helpers.R
\name{execute_sell}
\alias{execute_sell}
\title{Execute Sell Trade Function}
\usage{
execute_sell(sell, pobj, prices)
}
\arguments{
\item{sell}{data.frame with sell trade information. Result of get_sell_trades
function}

\item{pobj}{portfolio object}

\item{prices}{current prices of symbols being traded}
}
\value{
updated portfolio object
}
\description{
Updates portfolio with sell trade provided
}
