% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade-helpers.R
\name{execute_trade_pair}
\alias{execute_trade_pair}
\title{Execute Trade Pair Function}
\usage{
execute_trade_pair(buy, sell, pobj, prices, amount, lot_size = 1)
}
\arguments{
\item{buy}{buy symbol to buy}

\item{sell}{sell symbol}

\item{pobj}{portfolio object}

\item{prices}{current prices of symbols being traded}

\item{amount}{trade amount}

\item{lot_size}{min lot size}
}
\value{
updated portfolio object
}
\description{
Updates portfolio with trade pair provided
}
