% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{filter_constraints}
\alias{filter_constraints}
\title{Function to filter Constraints by index}
\usage{
filter_constraints(constraints, index)
}
\arguments{
\item{constraints}{constraints object}

\item{index}{numeric index to filter constraints by}
}
\description{
Function to filter Constraints by index
}
