% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{get_buy_trades}
\alias{get_buy_trades}
\alias{get_buy_trades.data.frame}
\alias{get_buy_trades.estimates}
\alias{get_buy_trades.portfolio}
\alias{get_buy_trades.character}
\title{Get Buy Trades}
\usage{
get_buy_trades(obj, symbols, amount, lot_size)

\method{get_buy_trades}{data.frame}(obj, symbols = NULL, amount,
  lot_size = 1)

\method{get_buy_trades}{estimates}(obj, symbols = NULL, amount,
  lot_size = 1)

\method{get_buy_trades}{portfolio}(obj, symbols = NULL, amount,
  lot_size = 1)

\method{get_buy_trades}{character}(obj, symbols = NULL, amount,
  lot_size = 1)
}
\arguments{
\item{obj}{object to use for buy trades}

\item{symbols}{vector of holding symbols to filter sell tickets by}

\item{amount}{trade amount}

\item{lot_size}{minimum share lot size}
}
\value{
data.frame with buy trades
}
\description{
Get Buy Trades
}
