% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{get_estimated_holdings_market_value}
\alias{get_estimated_holdings_market_value}
\title{Get Estimated Holdings Market Value}
\usage{
get_estimated_holdings_market_value(pobj, eobj)
}
\arguments{
\item{pobj}{portfolio object}

\item{eobj}{estimates object}
}
\value{
data.frame with updated holdings market values
}
\description{
Update portfolio's holding market values with mu estimates
}
