% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{get_estimated_port_stats}
\alias{get_estimated_port_stats}
\title{Get Estimated Porfolio Stats}
\usage{
get_estimated_port_stats(pobj, eobj, port_only = FALSE)
}
\arguments{
\item{pobj}{portfolio object}

\item{eobj}{estimates object}

\item{port_only}{logical option to return only portfolio levels stats}
}
\value{
tibble with estimate stats for portfolio and investments only
}
\description{
Function calculates estimated portfolio return, risk, sharpe and yield
statistics. Computes for both total portfolio and investments only
}
