% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{get_estimated_port_values}
\alias{get_estimated_port_values}
\title{Get Estimated Portfolio Values}
\usage{
get_estimated_port_values(pobj, eobj)
}
\arguments{
\item{pobj}{portfolio object}

\item{eboj}{estimates object}
}
\value{
data.frame with estimated portfolio values
}
\description{
Calculates estimated portfolio return, risk, sharpe and income in dollar
values
}
