% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{get_estimates_stats}
\alias{get_estimates_stats}
\title{Get Estimates Stats}
\usage{
get_estimates_stats(eobj)
}
\arguments{
\item{eobj}{estimates object}
}
\description{
Function gets Estimates stats by symobl
}
\details{
Returns mu return, standard deviation, sharpe ratio and annual dividend
}
