% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{get_mu}
\alias{get_mu}
\title{Get Mu Estimates from Estiamates Object}
\usage{
get_mu(eobj)
}
\arguments{
\item{eobj}{estimates object}
}
\value{
data.frame with ordered symbols
}
\description{
Get Mu Estimates from Estiamates Object
}
