% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{get_sell_trades}
\alias{get_sell_trades}
\title{Get Sell Trades}
\usage{
get_sell_trades(pobj, symbols, amount, lot_size = 1, partial = TRUE)
}
\arguments{
\item{pobj}{portfolio object}

\item{symbols}{vector of holding symbols to filter sell tickets by}

\item{amount}{trade amount}

\item{lot_size}{minimum share lot size}

\item{partial}{logical option to allow for partial trade tickets with an
amount less than the amount parameter value provided}
}
\value{
data.frame with possible sell trades
}
\description{
Returns all possible sell trades from a portfolio object
}
\details{
Checks current holdings and generates sell tickets based on symbol, trade
amount and lot_size
}
