% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{get_symbol_estimates_share}
\alias{get_symbol_estimates_share}
\title{Get Share of Total Portfolio By Symbol for Estimates}
\usage{
get_symbol_estimates_share(pobj, eobj)
}
\arguments{
\item{pobj}{portfolio object}

\item{eboj}{estimates object}
}
\value{
data.frame with portfolio share for all estimates symbol
}
\description{
Aggregates holdings portfolio share for all Estimate Symbols
}
