% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{group_constraint}
\alias{group_constraint}
\title{Group Constraint Constructer}
\usage{
group_constraint(symbols, min, max)
}
\arguments{
\item{symbols}{1 or more symbols}

\item{min}{minimum constraint value. inclusive}

\item{max}{maximum constraint value. inclusive}
}
\value{
group_constraint object
}
\description{
Group constraints constrain the total share of a portfolio's market value 2
or more symbols can have. Inherits from constraint class
}
