% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{meet_constraint}
\alias{meet_constraint}
\alias{meet_constraint.symbol_constraint}
\alias{meet_constraint.cash_constraint}
\alias{meet_constraint.cardinality_constraint}
\alias{meet_constraint.group_constraint}
\alias{meet_constraint.performance_constraint}
\title{Meet Constraint}
\usage{
meet_constraint(constraint, pobj, cobj, eobj, prices, trade_pairs,
  minimize, target, amount, lot_size, max_iter, ...)

\method{meet_constraint}{symbol_constraint}(constraint, pobj, cobj, eobj,
  prices, trade_pairs, target, minimize, amount, lot_size, max_iter = 5,
  ...)

\method{meet_constraint}{cash_constraint}(constraint, pobj, cobj, eobj,
  prices, trade_pairs, target, minimize, amount, lot_size, max_iter = 5,
  ...)

\method{meet_constraint}{cardinality_constraint}(constraint, pobj, cobj,
  eobj, prices, trade_pairs, target, minimize, amount, lot_size,
  max_iter = 5, ...)

\method{meet_constraint}{group_constraint}(constraint, pobj, cobj, eobj,
  prices, trade_pairs, target, minimize, amount, lot_size, max_iter = 5,
  ...)

\method{meet_constraint}{performance_constraint}(constraint, pobj, cobj,
  eobj, prices, trade_pairs, target, minimize, amount, lot_size,
  max_iter = 5, ...)
}
\arguments{
\item{constraint}{constraint object}

\item{pobj}{portfolio object}

\item{cobj}{constraints object}

\item{eobj}{estimates object}

\item{prices}{current symbol prices}

\item{trade_pairs}{possible trade pairs}

\item{minimize}{logical option to minimize target objective}

\item{target}{optimization target}

\item{amount}{trade amount for nbto}

\item{lot_size}{trade lot minimum size}

\item{max_iter}{maximum number of iterations for nbto}

\item{...}{additional parameters. not currently implemented}
}
\value{
data.frame with summary of constraint check
}
\description{
Checks portfolio against constraint and updates portfolio with nbto
}
