% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{nbto_optimize}
\alias{nbto_optimize}
\title{Optimize NBTO}
\usage{
nbto_optimize(pobj, cobj, eobj, prices, trade_pairs, n_pairs, target,
  minimize, amount, lot_size, max_iter = 5, max_runtime = 60,
  improve_lag = 5, min_improve = 0.001, include_port = TRUE,
  update_trade_pairs = TRUE, random_pairs = TRUE)
}
\arguments{
\item{pobj}{portfolio object}

\item{cobj}{constraints object}

\item{eobj}{estimates object}

\item{prices}{data.frame with current symbol prices. Has to contain all
symbols included in estimates object. Should also include dividend}

\item{trade_pairs}{trade pairs to consider. trades limited to the trade pairs
provided}

\item{target}{target objective}

\item{minimize}{logical flag to minimize target objective}

\item{amount}{trade amount in dollars}

\item{lot_size}{lot size for trades}

\item{include_port}{logical flag to include the portfolio object provided in
the canidate list. Default is FALSE.}

\item{update_trade_pairs}{logical option to update the trade pairs after
optimize step. Default is FALSE}

\item{random_pairs}{logical option to take weighted random sample of trades
or top_n}
}
\value{
list with list of portfolios, portfolio values, updated trade pairs
  and runtime
}
\description{
Function to iteratively apply NBTO function to portfolio. Includes runtime
and max iteration constraints. Used internally in optimize and meet
constraints functions
}
