% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{new_estimates}
\alias{new_estimates}
\title{Estimates Class Constructer function}
\usage{
new_estimates(symbols, start_date, end_date, grain, periods, prices,
  returns, mu, sigma, dividends)
}
\arguments{
\item{symbols}{vector of security symbols}

\item{start_date}{start date for historical prices}

\item{end_date}{end date for historical prices}

\item{grain}{time grain of returns. Can be either day, week, month, year}

\item{periods}{number of time periods}

\item{prices}{optional input for security prices}

\item{returns}{optional input for security returns}

\item{mu}{optional input for mean estimates}

\item{sigma}{optional input for sigma estimates}

\item{dividends}{optional input for dividend estimates}
}
\value{
estimates object
}
\description{
Creates an estimate object
}
