% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{optimize}
\alias{optimize}
\title{Optimize Portfolio Optimization}
\usage{
optimize(obj, n_pairs, amount, lot_size = 1, max_iter = 10,
  max_runtime = 300, improve_lag = 2, min_improve = 0.001,
  plot_iter = TRUE)
}
\arguments{
\item{obj}{portfolio optimization object to optimize}

\item{n_pairs}{number of trade pairs consider for each optimization step}

\item{amount}{trade amount}

\item{lot_size}{minimum share lot size}

\item{max_iter}{maximum number of iterations}

\item{max_runtime}{max runtime in seconds}

\item{improve_lag}{number of iterations lags to compare min improvement
against}

\item{min_improve}{minimum improvement of current iteration over improve_lag.
if not met, routine stopped}

\item{plot_iter}{logical option to plot interative results}
}
\value{
updated portfolio optimization object
}
\description{
Executes optimize routine on porfolio optimization object
}
