% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{performance_constraint}
\alias{performance_constraint}
\title{Performance Constraint Constructer}
\usage{
performance_constraint(statistic, min, max)
}
\arguments{
\item{statistic}{character input for portfolio statistic. valid statistics
are mu, sd, sharpe or yield}

\item{min}{minimum constraint value. inclusive}

\item{max}{maximum constraint value. inclusive}
}
\value{
performance_constraint object
}
\description{
Performance constraints constrain the value of a portfolio's estimated
statistics. Inherits from class constraint
}
