% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{po_constraints_charts}
\alias{po_constraints_charts}
\title{Portfolio Optimization Constraints Chart}
\usage{
po_constraints_charts(obj)
}
\arguments{
\item{obj}{portfolio optimization object}
}
\value{
ggplot object
}
\description{
Creates a ggplot chart of the portfolio optimization constraints values with
bounds for each optimization iteration
}
