% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{po_symbol_share_chart}
\alias{po_symbol_share_chart}
\title{Portfolio Optimization Symbol Share Chart}
\usage{
po_symbol_share_chart(obj)
}
\arguments{
\item{obj}{portfolio optimization object}
}
\value{
ggplot object
}
\description{
Creates a ggplot chart of the portfolio symbol share amount for each
optimization iteration
}
