% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{po_target_chart}
\alias{po_target_chart}
\title{Portfolio Optimization Target Chart}
\usage{
po_target_chart(obj)
}
\arguments{
\item{obj}{portfolio optimization object}
}
\value{
ggplot object
}
\description{
Creates a ggplot chart of the portfolio optimization target values for each
optimization iteration
}
