% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{portfolio_optimization}
\alias{portfolio_optimization}
\title{Portfolio Optimization Constructor Function}
\usage{
portfolio_optimization(portfolio, estimates, constraints, prices = NULL,
  target, desc = "", version = 1)
}
\arguments{
\item{portfolio}{portfolio object}

\item{estimates}{estimates object}

\item{prices}{current symbol prices}

\item{target}{target objective}

\item{desc}{optional meta-data description input}

\item{version}{optional input for version}
}
\value{
portfolio_optimization class
}
\description{
Create a portfolio optimization object. Portfolio optimization objects can be
optimized with the optimize functions.
}
