% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{remove_constraint}
\alias{remove_constraint}
\title{Remove Constraint from Constraints Object}
\usage{
remove_constraint(constraints, index, id = NULL)
}
\arguments{
\item{constraints}{constraints object}

\item{index}{constraint numeric index to remove. reference to position in
constraints list}

\item{id}{constraint id number to remove. default is NULL. If id not null,
will over-ride index argument}
}
\value{
updated constraints object
}
\description{
Remove constraint object from constraints list of contraints. Can reference
constraint by either index or id
}
