% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{restrict_trading}
\alias{restrict_trading}
\title{Restrict Trading Symbols}
\usage{
restrict_trading(cobj, symbols)
}
\arguments{
\item{cobj}{constraints object}

\item{symbols}{vector of symbols to restict trading}
}
\value{
updated constraints object
}
\description{
Restrict symbols from being traded in portfolio optimization. Removes any
symbol constraints as well
}
\details{
Typical use case is to restict certain portfolio holdings from being traded,
due to tax implications or lack of liquidity
}
