% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{set_buy_symbols}
\alias{set_buy_symbols}
\title{Set Buy Symbols}
\usage{
set_buy_symbols(cobj, symbols)
}
\arguments{
\item{cobj}{constraints object}

\item{symbols}{vector of symbols to restict trading}
}
\value{
updated constraints object
}
\description{
Function sets the symbols to buy in the portfolio optimization. Restricts the
buy symbols in the possible trade pairs
}
