% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{set_sell_symbols}
\alias{set_sell_symbols}
\title{Set Sell Symbols}
\usage{
set_sell_symbols(cobj, symbols)
}
\arguments{
\item{cobj}{constraints object}

\item{symbols}{vector of symbols to restict trading}
}
\value{
updated constraints object
}
\description{
Function sets the symbols to sell in portfolio optimization. Restricts the
sell symbols in the possible trade pairs
}
