% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{symbol_constraint}
\alias{symbol_constraint}
\title{Symbol Constraint Constructer}
\usage{
symbol_constraint(symbols, min, max)
}
\arguments{
\item{symbols}{0 or more symbols to constrain. If NULL, sets min and max
values to all symbols}

\item{min}{minimum constraint value. inclusive}

\item{max}{maximum constraint value. inclusive}
}
\value{
symbol oonstaints object
}
\description{
Inherits from constraint class
}
