% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{trade_pairs}
\alias{trade_pairs}
\title{Create Trade Pairs Function}
\usage{
trade_pairs(eobj, cobj, target)
}
\arguments{
\item{eobj}{estimates object}

\item{cobj}{constraints object}

\item{target}{target objective}
}
\value{
trade pairs tibble
}
\description{
Given estimates, and target creates an expanded grid of possible trade pairs
with expected target impact (delta)
}
