% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{add_max_risk}
\alias{add_max_risk}
\title{Add Maximum Risk Performance Constraint to Contraints Object}
\usage{
add_max_risk(constraints, max = NULL)
}
\arguments{
\item{constraints}{constraints object}

\item{max}{maximum risk value}
}
\value{
updated constraints object
}
\description{
Add Maximum Risk Performance Constraint to Contraints Object
}
