% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{add_min_yield}
\alias{add_min_yield}
\title{Add Minimum Yield Performance Constraint to Contraints Object}
\usage{
add_min_yield(constraints, min = NULL)
}
\arguments{
\item{constraints}{constraints object}

\item{min}{minumum yield value}
}
\value{
updated constraints object
}
\description{
Add Minimum Yield Performance Constraint to Contraints Object
}
