% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{portfolio_optimization}
\alias{portfolio_optimization}
\title{Portfolio Optimization Constructor Function}
\usage{
portfolio_optimization(pobj, eobj, cobj, prices = NULL, target,
  desc = "", version = 1, backend = "sequential")
}
\arguments{
\item{pobj}{portfolio object}

\item{eobj}{estimates object}

\item{cobj}{constraints object}

\item{prices}{current symbol prices}

\item{target}{target objective}

\item{desc}{optional meta-data description input}

\item{version}{optional input for version}

\item{backend}{future backend mode. `sequential` creates single threaded
execution. `multisession` creates parrallel backend}
}
\value{
portfolio_optimization class
}
\description{
Create a portfolio optimization object. Portfolio optimization objects can be
optimized with the optimize functions.
}
