% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{select_optimal_portfolio}
\alias{select_optimal_portfolio}
\title{Select Optimal Portfolio}
\usage{
select_optimal_portfolio(portfolios, estimates, target, criteria)
}
\arguments{
\item{portfolios}{list of portfolio objects}

\item{estimates}{estimates object}

\item{target}{target objective}

\item{criteria}{string value. set to minimize to select smallest target value}
}
\value{
optimal portfolio object
}
\description{
Selects optimal portfolio given list of canidate portfolios, estimates, and a
target objective
}
