% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{trade_pairs}
\alias{trade_pairs}
\title{Create Trade Pairs Function}
\usage{
trade_pairs(portfolio, estimates, target)
}
\arguments{
\item{portfolio}{portfolio object}

\item{estimates}{estimates object}

\item{target}{target objective}
}
\value{
trade pairs data.frame
}
\description{
Given portfolio, estimates, and target creates a data.frame of possible trade
pairs with expected target impact (delta)
}
